requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.10.4
itsdangerous==0.24
requests==2.6.0
six==1.9.0
